Moving screen metrics to a new API class #197
