+++ Xamarin.Essentials/DeviceDisplay/DeviceDisplay.netstandard.cs
@@ -1,0 +1,11 @@
+﻿namespace Xamarin.Essentials
+{
+    public static partial class DeviceDisplay
+    {
+        static ScreenMetrics GetScreenMetrics() => throw new NotImplementedInReferenceAssemblyException();
+
+        static void StartScreenMetricsListeners() => throw new NotImplementedInReferenceAssemblyException();
+
+        static void StopScreenMetricsListeners() => throw new NotImplementedInReferenceAssemblyException();
+    }
+}
