+++ Xamarin.Essentials/Xamarin.Essentials.csproj
@@ -40,7 +40,7 @@
     <DebugType>pdbonly</DebugType>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="mdoc" Version="5.6.3" PrivateAssets="All" />
+    <PackageReference Include="mdoc" Version="5.6.4" PrivateAssets="All" />
     <PackageReference Include="MSBuild.Sdk.Extras" Version="1.2.2" PrivateAssets="All" />
     <Compile Include="**\*.shared.cs" />
     <Compile Include="**\*.shared.*.cs" />
