+++ docs/en/Xamarin.Essentials/DeviceDisplay.xml
@@ -1,0 +1,58 @@
+<Type Name="DeviceDisplay" FullName="Xamarin.Essentials.DeviceDisplay">
+  <TypeSignature Language="C#" Value="public static class DeviceDisplay" />
+  <TypeSignature Language="ILAsm" Value=".class public auto ansi abstract sealed beforefieldinit DeviceDisplay extends System.Object" />
+  <TypeSignature Language="DocId" Value="T:Xamarin.Essentials.DeviceDisplay" />
+  <AssemblyInfo>
+    <AssemblyName>Xamarin.Essentials</AssemblyName>
+    <AssemblyVersion>1.0.0.0</AssemblyVersion>
+  </AssemblyInfo>
+  <Base>
+    <BaseTypeName>System.Object</BaseTypeName>
+  </Base>
+  <Interfaces />
+  <Docs>
+    <summary>Represents information about the device screen.</summary>
+    <remarks>
+      <para />
+    </remarks>
+  </Docs>
+  <Members>
+    <Member MemberName="ScreenMetrics">
+      <MemberSignature Language="C#" Value="public static Xamarin.Essentials.ScreenMetrics ScreenMetrics { get; }" />
+      <MemberSignature Language="ILAsm" Value=".property valuetype Xamarin.Essentials.ScreenMetrics ScreenMetrics" />
+      <MemberSignature Language="DocId" Value="P:Xamarin.Essentials.DeviceDisplay.ScreenMetrics" />
+      <MemberType>Property</MemberType>
+      <AssemblyInfo>
+        <AssemblyVersion>1.0.0.0</AssemblyVersion>
+        <AssemblyName>Xamarin.Essentials</AssemblyName>
+      </AssemblyInfo>
+      <ReturnValue>
+        <ReturnType>Xamarin.Essentials.ScreenMetrics</ReturnType>
+      </ReturnValue>
+      <Docs>
+        <summary>Gets the screen metrics of the device.</summary>
+        <value>The screen metrics.</value>
+        <remarks>
+          <para />
+        </remarks>
+      </Docs>
+    </Member>
+    <Member MemberName="ScreenMetricsChanaged">
+      <MemberSignature Language="C#" Value="public static event Xamarin.Essentials.ScreenMetricsChanagedEventHandler ScreenMetricsChanaged;" />
+      <MemberSignature Language="ILAsm" Value=".event class Xamarin.Essentials.ScreenMetricsChanagedEventHandler ScreenMetricsChanaged" />
+      <MemberSignature Language="DocId" Value="E:Xamarin.Essentials.DeviceDisplay.ScreenMetricsChanaged" />
+      <MemberType>Event</MemberType>
+      <AssemblyInfo>
+        <AssemblyVersion>1.0.0.0</AssemblyVersion>
+        <AssemblyName>Xamarin.Essentials</AssemblyName>
+      </AssemblyInfo>
+      <ReturnValue>
+        <ReturnType>Xamarin.Essentials.ScreenMetricsChanagedEventHandler</ReturnType>
+      </ReturnValue>
+      <Docs>
+        <summary>Event that is triggered when the screen matrics change.</summary>
+        <remarks>Such as screen rotation.</remarks>
+      </Docs>
+    </Member>
+  </Members>
+</Type>
